Revert "Fix 2047 ("TripleOscillator: Oscillators are getting out of sync") (#3145)"
This reverts commit 291942f6553f973f3cb59d0c10e618a56eb96ee5 in an
attempt to fix issue #3292.

The problem that's addressed is very elusive:
* It cannot be reproduced reliably.
* It seems to be more prone to appear in release builds.
* It only seems to appear when two or more instruments are mixed into
  one channel.
* If you solo a problematic Triple Osc it goes away.

The observations above seem to indicate that there's rather a problem
with the mixing (due to multithreading?) and that the changes made to
Triple Osc seem to trigger this deeper problem more quickly.

+++ include/Oscillator.h
@@ -71,7 +71,7 @@ public:
 			const IntModel * _mod_algo_model,
 			const float & _freq,
 			const float & _detuning,
-			const double & _phase_offset,
+			const float & _phase_offset,
 			const float & _volume,
 			Oscillator * _m_subOsc = NULL );
 	virtual ~Oscillator()
@@ -160,10 +160,10 @@ private:
 	const float & m_freq;
 	const float & m_detuning;
 	const float & m_volume;
-	const double & m_ext_phaseOffset;
+	const float & m_ext_phaseOffset;
 	Oscillator * m_subOsc;
-	double m_phaseOffset;
-	double m_phase;
+	float m_phaseOffset;
+	float m_phase;
 	const SampleBuffer * m_userWave;
